TASK#15: Refactoring 1. Moved screen related components into screen folders
