+++ project/src/components/films-list/video-preview/video-preview.tsx
