+++ project/src/components/header/header.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link} from 'react-router-dom';
-import User from '../user/user';
+import User from './user/user';
 
 type HeaderProps = {
   title?: string,
