+++ project/src/components/header/user/user.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Link, useHistory } from 'react-router-dom';
-import { AppRoute } from '../../constants';
-import { State } from '../../store/reducer';
+import { AppRoute } from '../../../constants';
+import { State } from '../../../store/reducer';
 import { connect, ConnectedProps } from 'react-redux';
-import { AsyncDispatch } from '../../types/action';
-import { logout } from '../../store/user/user-thunks';
+import { AsyncDispatch } from '../../../types/action';
+import { logout } from '../../../store/user/user-thunks';
 
 const mapStateToProps = (state: State) => ({
   authStatus: state.user.authStatus,
