+++ project/src/screens/film-page/components/film-tabs/film-tabs.tsx
@@ -3,7 +3,7 @@ import { TabList, Tabs, Tab, TabPanel } from 'react-tabs';
 import Overview from './overview';
 import Details from './details';
 import Reviews from './reviews';
-import { FilmProps } from '../../types/film-props';
+import { FilmProps } from '../../../../types/film-props';
 
 function FilmTabs(props: FilmProps): JSX.Element {
   const [tabIndex, setTabIndex] = useState(0);
