+++ project/src/screens/film-page/film-page.tsx
@@ -5,9 +5,9 @@ import { State } from '../../store/reducer';
 import { scrollToFilmTitle } from '../../utils/side-effects';
 import Header from '../../components/header/header';
 import Footer from '../../components/footer/footer';
-import FilmTabs from '../../components/film-tabs/film-tabs';
+import FilmTabs from './components/film-tabs/film-tabs';
 import NotFound from '../not-found/not-found';
-import RelatedFilms from '../../components/related-films/related-films';
+import RelatedFilms from './components/related-films/related-films';
 import Spinner from '../../components/spinner/Spinner';
 import { AuthStatus } from '../../constants';
 import { useLoadFilm } from '../../hooks/films';
