+++ project/src/screens/main/components/genres/genres.tsx
